+++ src/core/cartridges/cartridges_str.rs
@@ -12,7 +12,7 @@ impl CartridgeBase<Rule, &str> for Cartridge<Rule> {
         )
     }
 
-    fn get_id(&self) -> i64 {
+    fn get_id(&self) -> i32 {
         self.id
     }
